Revamp window_demon, get connection_window running.
Also update debug to log to debug.log.
Update connection_entry to be a simple struct instead of a class, use it
to communicate to demon what it should print, not sure if it should be
used for anything else. A little scuffed as we transfer from
temp_connections.

Still wondering how to handle communication between model_angel and
window_demon. Can't initialize both with references to each other, so
maybe just pointers?