Add dns caching! What a piece of garbage!
Quickly wrote some code for dns caching that is *rather* bad.
Uses <map> because I didn't want to write a hashing function for
<unordered_map>. Also doesn't check for stale DNS records!

But it is no longer slow, and therefore objectively better.
Objectively.
Obviously.

It's also *encapsulated* garbage, so maybe I will never have to touch it
again. Probably not though.